+++ packages/cli/binding/src/cli/resolver.rs
@@ -1,0 +1,365 @@
+use std::{env, ffi::OsStr, iter, sync::Arc};
+
+use rustc_hash::FxHashMap;
+use vite_path::AbsolutePath;
+use vite_str::Str;
+use vite_task::config::user::{
+    AutoInput, EnabledCacheConfig, GlobWithBase, InputBase, UserCacheConfig, UserInputEntry,
+};
+
+use super::{
+    help::should_prepend_vitest_run,
+    types::{CliOptions, ResolvedSubcommand, ResolvedUniversalViteConfig, SynthesizableSubcommand},
+};
+
+/// Resolves synthesizable subcommands to concrete programs and arguments.
+/// Used by both direct CLI execution and CommandHandler.
+pub struct SubcommandResolver {
+    cli_options: Option<CliOptions>,
+    workspace_path: Arc<AbsolutePath>,
+}
+
+impl std::fmt::Debug for SubcommandResolver {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("SubcommandResolver")
+            .field("has_cli_options", &self.cli_options.is_some())
+            .field("workspace_path", &self.workspace_path)
+            .finish()
+    }
+}
+
+impl SubcommandResolver {
+    pub fn new(workspace_path: Arc<AbsolutePath>) -> Self {
+        Self { cli_options: None, workspace_path }
+    }
+
+    pub fn with_cli_options(mut self, cli_options: CliOptions) -> Self {
+        self.cli_options = Some(cli_options);
+        self
+    }
+
+    fn cli_options(&self) -> anyhow::Result<&CliOptions> {
+        self.cli_options
+            .as_ref()
+            .ok_or_else(|| anyhow::anyhow!("CLI options not available (running without NAPI?)"))
+    }
+
+    pub(crate) async fn resolve_universal_vite_config(
+        &self,
+    ) -> anyhow::Result<ResolvedUniversalViteConfig> {
+        let cli_options = self.cli_options()?;
+        let workspace_path_str = self
+            .workspace_path
+            .as_path()
+            .to_str()
+            .ok_or_else(|| anyhow::anyhow!("workspace path is not valid UTF-8"))?;
+        let vite_config_json =
+            (cli_options.resolve_universal_vite_config)(workspace_path_str.to_string()).await?;
+
+        Ok(serde_json::from_str(&vite_config_json).inspect_err(|_| {
+            tracing::error!("Failed to parse vite config: {vite_config_json}");
+        })?)
+    }
+
+    /// Resolve a synthesizable subcommand to a concrete program, args, cache config, and envs.
+    pub(super) async fn resolve(
+        &self,
+        subcommand: SynthesizableSubcommand,
+        resolved_vite_config: Option<&ResolvedUniversalViteConfig>,
+        envs: &Arc<FxHashMap<Arc<OsStr>, Arc<OsStr>>>,
+        cwd: &Arc<AbsolutePath>,
+    ) -> anyhow::Result<ResolvedSubcommand> {
+        match subcommand {
+            SynthesizableSubcommand::Lint { mut args } => {
+                let cli_options = self.cli_options()?;
+                let resolved = (cli_options.lint)().await?;
+                let js_path = resolved.bin_path;
+                let js_path_str = js_path
+                    .to_str()
+                    .ok_or_else(|| anyhow::anyhow!("lint JS path is not valid UTF-8"))?;
+                let owned_resolved_vite_config;
+                let resolved_vite_config = if let Some(config) = resolved_vite_config {
+                    config
+                } else {
+                    owned_resolved_vite_config = self.resolve_universal_vite_config().await?;
+                    &owned_resolved_vite_config
+                };
+
+                if let (Some(_), Some(config_file)) =
+                    (&resolved_vite_config.lint, &resolved_vite_config.config_file)
+                {
+                    args.insert(0, "-c".to_string());
+                    args.insert(1, config_file.clone());
+                }
+
+                Ok(ResolvedSubcommand {
+                    program: Arc::from(OsStr::new("node")),
+                    args: iter::once(Str::from("--disable-warning=MODULE_TYPELESS_PACKAGE_JSON"))
+                        .chain(iter::once(Str::from(js_path_str)))
+                        .chain(args.into_iter().map(Str::from))
+                        .collect(),
+                    cache_config: UserCacheConfig::with_config(EnabledCacheConfig {
+                        env: Some(Box::new([Str::from("OXLINT_TSGOLINT_PATH")])),
+                        untracked_env: None,
+                        input: None,
+                    }),
+                    envs: merge_resolved_envs(envs, resolved.envs),
+                })
+            }
+            SynthesizableSubcommand::Fmt { mut args } => {
+                let cli_options = self.cli_options()?;
+                let resolved = (cli_options.fmt)().await?;
+                let js_path = resolved.bin_path;
+                let js_path_str = js_path
+                    .to_str()
+                    .ok_or_else(|| anyhow::anyhow!("fmt JS path is not valid UTF-8"))?;
+                let owned_resolved_vite_config;
+                let resolved_vite_config = if let Some(config) = resolved_vite_config {
+                    config
+                } else {
+                    owned_resolved_vite_config = self.resolve_universal_vite_config().await?;
+                    &owned_resolved_vite_config
+                };
+
+                if let (Some(_), Some(config_file)) =
+                    (&resolved_vite_config.fmt, &resolved_vite_config.config_file)
+                {
+                    args.insert(0, "-c".to_string());
+                    args.insert(1, config_file.clone());
+                }
+
+                Ok(ResolvedSubcommand {
+                    program: Arc::from(OsStr::new("node")),
+                    args: iter::once(Str::from(js_path_str))
+                        .chain(args.into_iter().map(Str::from))
+                        .collect(),
+                    cache_config: UserCacheConfig::with_config(EnabledCacheConfig {
+                        env: None,
+                        untracked_env: None,
+                        input: None,
+                    }),
+                    envs: merge_resolved_envs(envs, resolved.envs),
+                })
+            }
+            SynthesizableSubcommand::Build { args } => {
+                let cli_options = self.cli_options()?;
+                let resolved = (cli_options.vite)().await?;
+                let js_path = resolved.bin_path;
+                let js_path_str = js_path
+                    .to_str()
+                    .ok_or_else(|| anyhow::anyhow!("vite JS path is not valid UTF-8"))?;
+
+                Ok(ResolvedSubcommand {
+                    program: Arc::from(OsStr::new("node")),
+                    args: iter::once(Str::from(js_path_str))
+                        .chain(iter::once(Str::from("build")))
+                        .chain(args.into_iter().map(Str::from))
+                        .collect(),
+                    cache_config: UserCacheConfig::with_config(EnabledCacheConfig {
+                        env: Some(Box::new([Str::from("VITE_*")])),
+                        untracked_env: None,
+                        input: Some(build_pack_cache_inputs()),
+                    }),
+                    envs: merge_resolved_envs_with_version(envs, resolved.envs),
+                })
+            }
+            SynthesizableSubcommand::Test { args } => {
+                let cli_options = self.cli_options()?;
+                let resolved = (cli_options.test)().await?;
+                let js_path = resolved.bin_path;
+                let js_path_str = js_path
+                    .to_str()
+                    .ok_or_else(|| anyhow::anyhow!("test JS path is not valid UTF-8"))?;
+                let prepend_run = should_prepend_vitest_run(&args);
+                let vitest_args: Vec<Str> = if prepend_run {
+                    iter::once(Str::from("run")).chain(args.into_iter().map(Str::from)).collect()
+                } else {
+                    args.into_iter().map(Str::from).collect()
+                };
+
+                Ok(ResolvedSubcommand {
+                    program: Arc::from(OsStr::new("node")),
+                    args: iter::once(Str::from(js_path_str)).chain(vitest_args).collect(),
+                    cache_config: UserCacheConfig::with_config(EnabledCacheConfig {
+                        env: None,
+                        untracked_env: None,
+                        input: Some(vec![
+                            UserInputEntry::Auto(AutoInput { auto: true }),
+                            exclude_glob("!node_modules/.vite-temp/**", InputBase::Package),
+                            exclude_glob(
+                                "!node_modules/.vite/vitest/**/results.json",
+                                InputBase::Package,
+                            ),
+                        ]),
+                    }),
+                    envs: merge_resolved_envs_with_version(envs, resolved.envs),
+                })
+            }
+            SynthesizableSubcommand::Pack { args } => {
+                let cli_options = self.cli_options()?;
+                let resolved = (cli_options.pack)().await?;
+                let js_path = resolved.bin_path;
+                let js_path_str = js_path
+                    .to_str()
+                    .ok_or_else(|| anyhow::anyhow!("pack JS path is not valid UTF-8"))?;
+
+                Ok(ResolvedSubcommand {
+                    program: Arc::from(OsStr::new("node")),
+                    args: iter::once(Str::from(js_path_str))
+                        .chain(args.into_iter().map(Str::from))
+                        .collect(),
+                    cache_config: UserCacheConfig::with_config(EnabledCacheConfig {
+                        env: None,
+                        untracked_env: None,
+                        input: Some(build_pack_cache_inputs()),
+                    }),
+                    envs: merge_resolved_envs(envs, resolved.envs),
+                })
+            }
+            SynthesizableSubcommand::Dev { args } => {
+                let cli_options = self.cli_options()?;
+                let resolved = (cli_options.vite)().await?;
+                let js_path = resolved.bin_path;
+                let js_path_str = js_path
+                    .to_str()
+                    .ok_or_else(|| anyhow::anyhow!("vite JS path is not valid UTF-8"))?;
+
+                Ok(ResolvedSubcommand {
+                    program: Arc::from(OsStr::new("node")),
+                    args: iter::once(Str::from(js_path_str))
+                        .chain(iter::once(Str::from("dev")))
+                        .chain(args.into_iter().map(Str::from))
+                        .collect(),
+                    cache_config: UserCacheConfig::disabled(),
+                    envs: merge_resolved_envs_with_version(envs, resolved.envs),
+                })
+            }
+            SynthesizableSubcommand::Preview { args } => {
+                let cli_options = self.cli_options()?;
+                let resolved = (cli_options.vite)().await?;
+                let js_path = resolved.bin_path;
+                let js_path_str = js_path
+                    .to_str()
+                    .ok_or_else(|| anyhow::anyhow!("vite JS path is not valid UTF-8"))?;
+
+                Ok(ResolvedSubcommand {
+                    program: Arc::from(OsStr::new("node")),
+                    args: iter::once(Str::from(js_path_str))
+                        .chain(iter::once(Str::from("preview")))
+                        .chain(args.into_iter().map(Str::from))
+                        .collect(),
+                    cache_config: UserCacheConfig::disabled(),
+                    envs: merge_resolved_envs_with_version(envs, resolved.envs),
+                })
+            }
+            SynthesizableSubcommand::Doc { args } => {
+                let cli_options = self.cli_options()?;
+                let resolved = (cli_options.doc)().await?;
+                let js_path = resolved.bin_path;
+                let js_path_str = js_path
+                    .to_str()
+                    .ok_or_else(|| anyhow::anyhow!("doc JS path is not valid UTF-8"))?;
+
+                Ok(ResolvedSubcommand {
+                    program: Arc::from(OsStr::new("node")),
+                    args: iter::once(Str::from(js_path_str))
+                        .chain(args.into_iter().map(Str::from))
+                        .collect(),
+                    cache_config: UserCacheConfig::with_config(EnabledCacheConfig {
+                        env: None,
+                        untracked_env: None,
+                        input: None,
+                    }),
+                    envs: merge_resolved_envs(envs, resolved.envs),
+                })
+            }
+            SynthesizableSubcommand::Check { .. } => {
+                anyhow::bail!(
+                    "Check is a composite command and cannot be resolved to a single subcommand"
+                );
+            }
+            SynthesizableSubcommand::Install { args } => {
+                let package_manager =
+                    vite_install::PackageManager::builder(cwd).build_with_default().await?;
+                let resolve_command = package_manager.resolve_install_command(&args);
+
+                let merged_envs = {
+                    let mut env_map = FxHashMap::clone(envs);
+                    for (k, v) in resolve_command.envs {
+                        env_map.insert(Arc::from(OsStr::new(&k)), Arc::from(OsStr::new(&v)));
+                    }
+                    Arc::new(env_map)
+                };
+
+                Ok(ResolvedSubcommand {
+                    program: Arc::<OsStr>::from(
+                        OsStr::new(&resolve_command.bin_path).to_os_string(),
+                    ),
+                    args: resolve_command.args.into_iter().map(Str::from).collect(),
+                    cache_config: UserCacheConfig::with_config(EnabledCacheConfig {
+                        env: None,
+                        untracked_env: None,
+                        input: None,
+                    }),
+                    envs: merged_envs,
+                })
+            }
+        }
+    }
+}
+
+/// Create a negative glob entry to exclude a pattern from cache fingerprinting.
+fn exclude_glob(pattern: &str, base: InputBase) -> UserInputEntry {
+    UserInputEntry::GlobWithBase(GlobWithBase { pattern: Str::from(pattern), base })
+}
+
+/// Common cache input entries for build/pack commands.
+/// Excludes .vite-temp config files and dist output files that are both read and written.
+/// TODO: The hardcoded `!dist/**` exclusion is a temporary workaround. It will be replaced
+/// by a runner-aware approach that automatically excludes task output directories.
+fn build_pack_cache_inputs() -> Vec<UserInputEntry> {
+    vec![
+        UserInputEntry::Auto(AutoInput { auto: true }),
+        exclude_glob("!node_modules/.vite-temp/**", InputBase::Workspace),
+        exclude_glob("!node_modules/.vite-temp/**", InputBase::Package),
+        exclude_glob("!dist/**", InputBase::Package),
+    ]
+}
+
+/// Cache input entries for the check command.
+/// The vp check subprocess is a full vp CLI process (not resolved to a binary like
+/// build/lint/fmt), so it accesses additional directories that must be excluded:
+/// - `.vite-temp`: config compilation cache, read+written during vp CLI startup
+/// - `.vite/task-cache`: task runner state files that change after each run
+pub(super) fn check_cache_inputs() -> Vec<UserInputEntry> {
+    vec![
+        UserInputEntry::Auto(AutoInput { auto: true }),
+        exclude_glob("!node_modules/.vite-temp/**", InputBase::Workspace),
+        exclude_glob("!node_modules/.vite-temp/**", InputBase::Package),
+        exclude_glob("!node_modules/.vite/task-cache/**", InputBase::Workspace),
+        exclude_glob("!node_modules/.vite/task-cache/**", InputBase::Package),
+    ]
+}
+
+fn merge_resolved_envs(
+    envs: &Arc<FxHashMap<Arc<OsStr>, Arc<OsStr>>>,
+    resolved_envs: Vec<(String, String)>,
+) -> Arc<FxHashMap<Arc<OsStr>, Arc<OsStr>>> {
+    let mut envs = FxHashMap::clone(envs);
+    for (k, v) in resolved_envs {
+        envs.entry(Arc::from(OsStr::new(&k))).or_insert_with(|| Arc::from(OsStr::new(&v)));
+    }
+    Arc::new(envs)
+}
+
+/// Merge resolved envs and inject VP_VERSION for rolldown-vite branding.
+fn merge_resolved_envs_with_version(
+    envs: &Arc<FxHashMap<Arc<OsStr>, Arc<OsStr>>>,
+    resolved_envs: Vec<(String, String)>,
+) -> Arc<FxHashMap<Arc<OsStr>, Arc<OsStr>>> {
+    let mut merged = merge_resolved_envs(envs, resolved_envs);
+    let map = Arc::make_mut(&mut merged);
+    map.entry(Arc::from(OsStr::new("VP_VERSION")))
+        .or_insert_with(|| Arc::from(OsStr::new(env!("CARGO_PKG_VERSION"))));
+    merged
+}
